Correct version of the api, I know it is probably a little too late
In preparation for our meeting in a couple of hours, I started revealing what we already have on GitHub. It was only then I realised the files I uploaded, including the database schemer are incorrect. It appears I got mixed up with folders on my machine. This is in fact the 100% api I created. My appologies.