registration feature
The registration is now working. I am not yet certain what the error was, but I have a sspicion. It only started working after I manually recreated the table and wrote out the variables by hand instead of copying and pasting whilst shifting. Thus I suspect there was a typo either in the code or the sql file. The original order of username, type, email, and password is still in use.