Update userclass.php
register method has been updated to check if the email already exists. As commented in the code, I believe the ->execute function returns boolean true on whether the data has been entered. Therefore it can be used to redirect successful registerers to the login page.
I am experiencing problems with the login feature, thus it is still under development.