Shifts changes
The functions have been updated as requested.
I created a created time and status updated time columns in shifts, and request time and approval time columns in requests.
We could either have the approval time in requests as the start time of a shift, and the status update time as the end time. Alternatively, I could create a separate column to show start time in shifts, but that means we would have to come up with a functionality for the admin to indicate the task has been started.
I have opted to get times and dates using php functions rather than having timestamp column type in sql. I feel php gives me more options, especially in forcing timezones.
The date and time are in one column. I believe if we ever need to perform calculations with them, string functions can be used to extract either the date or time on its own.