shift changes
I added changes to the request function of the shifts class. The function now checks if the request has already been made. Date and time is also inserted.
I changed the approved from default none to default false on the database itself. The sql file has been updated accordingly.
The requests.php is currently the dummy file for creating requests.